+++ clap_complete/tests/snapshots/sub_subcommands.bash
@@ -17,40 +17,40 @@ _my-app() {
                 cmd="my__app"
                 ;;
             my__app,help)
-                cmd="my__app__help"
+                cmd="my__app_subcmd_help"
                 ;;
             my__app,some_cmd)
-                cmd="my__app__some_cmd"
+                cmd="my__app_subcmd_some_cmd"
                 ;;
             my__app,some_cmd_alias)
-                cmd="my__app__some_cmd"
+                cmd="my__app_subcmd_some_cmd"
                 ;;
             my__app,test)
-                cmd="my__app__test"
+                cmd="my__app_subcmd_test"
                 ;;
-            my__app__help,help)
-                cmd="my__app__help__help"
+            my__app_subcmd_help,help)
+                cmd="my__app_subcmd_help_subcmd_help"
                 ;;
-            my__app__help,some_cmd)
-                cmd="my__app__help__some_cmd"
+            my__app_subcmd_help,some_cmd)
+                cmd="my__app_subcmd_help_subcmd_some_cmd"
                 ;;
-            my__app__help,test)
-                cmd="my__app__help__test"
+            my__app_subcmd_help,test)
+                cmd="my__app_subcmd_help_subcmd_test"
                 ;;
-            my__app__help__some_cmd,sub_cmd)
-                cmd="my__app__help__some_cmd__sub_cmd"
+            my__app_subcmd_help_subcmd_some_cmd,sub_cmd)
+                cmd="my__app_subcmd_help_subcmd_some_cmd_subcmd_sub_cmd"
                 ;;
-            my__app__some_cmd,help)
-                cmd="my__app__some_cmd__help"
+            my__app_subcmd_some_cmd,help)
+                cmd="my__app_subcmd_some_cmd_subcmd_help"
                 ;;
-            my__app__some_cmd,sub_cmd)
-                cmd="my__app__some_cmd__sub_cmd"
+            my__app_subcmd_some_cmd,sub_cmd)
+                cmd="my__app_subcmd_some_cmd_subcmd_sub_cmd"
                 ;;
-            my__app__some_cmd__help,help)
-                cmd="my__app__some_cmd__help__help"
+            my__app_subcmd_some_cmd_subcmd_help,help)
+                cmd="my__app_subcmd_some_cmd_subcmd_help_subcmd_help"
                 ;;
-            my__app__some_cmd__help,sub_cmd)
-                cmd="my__app__some_cmd__help__sub_cmd"
+            my__app_subcmd_some_cmd_subcmd_help,sub_cmd)
+                cmd="my__app_subcmd_some_cmd_subcmd_help_subcmd_sub_cmd"
                 ;;
             *)
                 ;;
@@ -72,7 +72,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__help)
+        my__app_subcmd_help)
             opts="test some_cmd help"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -86,7 +86,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__help__help)
+        my__app_subcmd_help_subcmd_help)
             opts=""
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -100,7 +100,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__help__some_cmd)
+        my__app_subcmd_help_subcmd_some_cmd)
             opts="sub_cmd"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -114,7 +114,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__help__some_cmd__sub_cmd)
+        my__app_subcmd_help_subcmd_some_cmd_subcmd_sub_cmd)
             opts=""
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 4 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -128,7 +128,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__help__test)
+        my__app_subcmd_help_subcmd_test)
             opts=""
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -142,7 +142,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__some_cmd)
+        my__app_subcmd_some_cmd)
             opts="-h -V --help --version sub_cmd help"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -156,7 +156,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__some_cmd__help)
+        my__app_subcmd_some_cmd_subcmd_help)
             opts="sub_cmd help"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -170,7 +170,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__some_cmd__help__help)
+        my__app_subcmd_some_cmd_subcmd_help_subcmd_help)
             opts=""
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 4 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -184,7 +184,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__some_cmd__help__sub_cmd)
+        my__app_subcmd_some_cmd_subcmd_help_subcmd_sub_cmd)
             opts=""
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 4 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -198,7 +198,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__some_cmd__sub_cmd)
+        my__app_subcmd_some_cmd_subcmd_sub_cmd)
             opts="-h -V --config --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
@@ -216,7 +216,7 @@ _my-app() {
             COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
             return 0
             ;;
-        my__app__test)
+        my__app_subcmd_test)
             opts="-h -V --case --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 2 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )